+++ NvimView/NvimView/NvimView+Resize.swift
@@ -52,8 +52,12 @@ extension NvimView {
       return
     }
 
-    self.offset.x = floor((size.width - self.cellSize.width * CGFloat(discreteSize.width)) / 2)
-    self.offset.y = floor((size.height - self.cellSize.height * CGFloat(discreteSize.height)) / 2)
+    self.offset.x = floor(
+      (size.width - self.cellSize.width * CGFloat(discreteSize.width)) / 2
+    )
+    self.offset.y = floor(
+      (size.height - self.cellSize.height * CGFloat(discreteSize.height)) / 2
+    )
 
     self.bridge
       .resize(width: discreteSize.width, height: discreteSize.height)
@@ -62,32 +66,50 @@ extension NvimView {
 
   private func launchNeoVim(_ size: Size) {
     logger.info("=== Starting neovim...")
-    let sockPath = URL(fileURLWithPath: NSTemporaryDirectory()).appendingPathComponent("vimr_\(self.uuid).sock").path
+    let sockPath = URL(
+      fileURLWithPath: NSTemporaryDirectory()
+    ).appendingPathComponent("vimr_\(self.uuid).sock").path
 
     self.api.msgpackRawStream
-        .subscribe(onNext: { msg in
-          switch msg {
-          case let .notification(method, params):
-            print("NOTIFICATION: \(method) with \(params.count) elements")
-          case let .error(_, msg):
-            print("MSG ERROR: \(msg)")
-          default:
-            print("???")
-            break
-          }
-        }, onError: { print("ERROR: \($0)" )})
-        .disposed(by: self.disposeBag)
+      .subscribe(onNext: { msg in
+        switch msg {
 
-    // We wait here, since the user of NvimView cannot subscribe on the Completable. We could demand that the user
-    // call launchNeoVim() by themselves, but...
+        case let .notification(method, params):
+          logger.debug("NOTIFICATION: \(method): \(params)")
+
+          guard method == NvimView.rpcEventName else { return }
+          self.eventsSubject.onNext(.rpcEvent(method, params))
+
+        case let .error(_, msg):
+          logger.debug("MSG ERROR: \(msg)")
+
+        default:
+          logger.debug("???: This should not happen")
+          break
+
+        }
+      }, onError: { print("ERROR: \($0)") })
+      .disposed(by: self.disposeBag)
+
+    // We wait here, since the user of NvimView cannot subscribe
+    // on the Completable. We could demand that the user call launchNeoVim()
+    // by themselves, but...
     try? self.bridge
       .runLocalServerAndNvim(width: size.width, height: size.height)
       .andThen(self.api.run(at: sockPath))
+      .andThen(
+        self.sourceFileUrls.reduce(Completable.empty()) { prev, url in
+          prev.andThen(self.api
+                         .commandOutput(str: "source \(url.path)")
+                         .asCompletable())
+        }
+      )
+      .andThen(self.api.subscribe(event: NvimView.rpcEventName))
       .wait()
   }
 
   private func randomEmoji() -> String {
-    let idx = Int(arc4random_uniform(UInt32(emojis.count)))
+    let idx = Int.random(in: 0..<emojis.count)
     guard let scalar = UnicodeScalar(emojis[idx]) else {
       return "😎"
     }