+++ resources/release-notes.md
@@ -1,6 +1,8 @@
 # 0.26.0-???
 
 * GH-314: You can customize the key shortcut for all menu items in the *Shortcut* preferences pane.
+* GH-501: Add key shortcuts to toggle the Buffer List, Markdown Preview, and HTML Preview tools.
+* GH-649: Add commands that can control some of GUI elements.
 * Draw the disclosure triangle in appropriate color of the current color scheme (and improve handling of changes of `cwd` in the file browser).
 * ...